# Conflicts:
#	fe/fe-core/src/main/java/org/apache/doris/nereids/trees/plans/commands/insert/InsertIntoTableCommand.java

+++ be/src/olap/rowset/segment_v2/inverted_index_desc.h
@@ -31,9 +31,9 @@ public:
     static const std::string segment_suffix;
     static const std::string index_suffix;
     static const std::string index_name_separator;
-    static std::string get_temporary_index_path(const std::string& segment_path, uint32_t uuid,
+    static std::string get_temporary_index_path(const std::string& segment_path, uint64_t uuid,
                                                 const std::string& index_suffix_path);
-    static std::string get_index_file_name(const std::string& path, uint32_t uuid,
+    static std::string get_index_file_name(const std::string& path, uint64_t uuid,
                                            const std::string& index_suffix_path);
     static std::string get_index_file_name(const std::string& path);
     static const std::string get_temporary_null_bitmap_file_name() { return "null_bitmap"; }
